+++ lib/podcast_index.rb
@@ -4,14 +4,16 @@ require "active_support/core_ext/string/inflections"
 require "json"
 
 require_relative "podcast_index/api/request"
-require_relative "podcast_index/api/podcasts"
-require_relative "podcast_index/api/episodes"
-require_relative "podcast_index/api/search"
-require_relative "podcast_index/api/recent"
-require_relative "podcast_index/api/value"
+
 require_relative "podcast_index/api/categories"
-require_relative "podcast_index/podcast"
+require_relative "podcast_index/api/episodes"
+require_relative "podcast_index/api/podcasts"
+require_relative "podcast_index/api/recent"
+require_relative "podcast_index/api/search"
+require_relative "podcast_index/api/value"
+require_relative "podcast_index/category"
 require_relative "podcast_index/episode"
+require_relative "podcast_index/podcast"
 require_relative "podcast_index/soundbite"
 require_relative "podcast_index/value"
 
@@ -22,6 +24,7 @@ module PodcastIndex
 
   class Error < StandardError; end
   class PodcastNotFound < Error; end
+  class CategoryNotFound < Error; end
 
   def self.configure
     self.base_url = "https://api.podcastindex.org/api/1.0".freeze