+++ spec/podcast_index/api/categories_spec.rb
@@ -5,8 +5,7 @@ RSpec.describe PodcastIndex::Api::Categories do
     let(:fixture) { file_fixture("categories/list_response.json").read }
 
     before do
-      stub_request(:get, %r{/categories/list})
-        .to_return(body: fixture, status: 200)
+      stub_request(:get, %r{/categories/list}).to_return(body: fixture, status: 200)
     end
 
     it "returns the body of the response" do